segment id uint64

+++ @MF3D/ease_initialization.m
@@ -182,7 +182,7 @@ if clear_results || isempty(obj.A)
     obj.C_raw = zeros(K_new, T);
     obj.S = zeros(K_new, T);
     obj.labels = zeros(K_new,1);
-    obj.ids = zeros(K_new, 1);
+    obj.ids = uint64(zeros(K_new, 1));
     obj.match_status.status = zeros(1, K_new);
     obj.match_status.em_ids = cell(1, K_new);
     obj.match_status.confidence = zeros(1, K_new);
